+++ frontend/src/features/vessel_sidebar/actions/track_request/TrackRequest.js
@@ -1,4 +1,4 @@
-import { useCallback, useMemo } from 'react'
+import { useCallback, useMemo, useState, useEffect } from 'react'
 import TrackDepth from './TrackDepth'
 import DateRange from './DateRange'
 import { COLORS } from '../../../../constants/constants'
@@ -12,25 +12,27 @@ import { useDispatch, useSelector } from 'react-redux'
 import { MapComponentStyle } from '../../../commonStyles/MapComponent.style'
 import PositionsTable from './PositionsTable'
 import modifyVesselTrackDepth from '../../../../domain/use_cases/vessel/modifyVesselTrackDepth'
+import ExportTrack from './ExportTrack'
 
 /**
  * @typedef {object} TrackRequestProps
- * @property {boolean} isRightMenuOpen
- * @property {boolean} isTrackRequestOpen
- * @property {(isOpen: boolean) => void} setIsTrackRequestOpen
+ * @property {boolean} sidebarIsOpen
  */
 
 /**
  * @param {TrackRequestProps} props
  */
-const TrackRequest = ({ isRightMenuOpen, setIsTrackRequestOpen, isTrackRequestOpen }) => {
+const TrackRequest = ({ sidebarIsOpen }) => {
   const dispatch = useDispatch()
   /** @type {{ healthcheckTextWarning: string }} */
   const { healthcheckTextWarning } = useSelector(state => state.global)
+  /** @type {{ rightMenuIsOpen: boolean }} */
+  const { rightMenuIsOpen } = useSelector(state => state.global)
   /** @type {{ selectedVesselCustomTrackRequest: VesselNS.TrackRequest }} */
   const { selectedVesselCustomTrackRequest } = useSelector(state => state.vessel)
   /** @type {{ selectedVesselIdentity: VesselNS.VesselIdentity }} */
   const { selectedVesselIdentity } = useSelector(state => state.vessel)
+  const [trackRequestIsOpen, setTrackRequestIsOpen] = useState(false)
 
   /** @type {[Date, Date] | undefined} */
   const selectedVesselCustomDateRange = useMemo(
@@ -39,6 +41,12 @@ const TrackRequest = ({ isRightMenuOpen, setIsTrackRequestOpen, isTrackRequestOp
       : undefined,
     [selectedVesselCustomTrackRequest]
   )
+
+  useEffect(() => {
+    if (!sidebarIsOpen) {
+      setTrackRequestIsOpen(false)
+    }
+  }, [sidebarIsOpen])
 
   /**
    * @param {[Date, Date]} dateRange
@@ -73,8 +81,9 @@ const TrackRequest = ({ isRightMenuOpen, setIsTrackRequestOpen, isTrackRequestOp
   return (
     <>
       <TrackRequestButton
-        data-cy={'vessel-track-depth-selection'}
         healthcheckTextWarning={healthcheckTextWarning}
+        data-cy={'vessel-track-depth-selection'}
+        sidebarIsOpen={sidebarIsOpen}
         isRightMenuOpen={isRightMenuOpen}
         isTrackRequestOpen={isTrackRequestOpen}
         onClick={() => setIsTrackRequestOpen(!isTrackRequestOpen)}
@@ -84,6 +93,7 @@ const TrackRequest = ({ isRightMenuOpen, setIsTrackRequestOpen, isTrackRequestOp
       </TrackRequestButton>
       <TrackRequestBody
         healthcheckTextWarning={healthcheckTextWarning}
+        sidebarIsOpen={sidebarIsOpen}
         isRightMenuOpen={isRightMenuOpen}
         isTrackRequestOpen={isTrackRequestOpen}
       >
@@ -98,6 +108,7 @@ const TrackRequest = ({ isRightMenuOpen, setIsTrackRequestOpen, isTrackRequestOp
           onChange={updateDateRange}
           placeholder={'Choisir une période précise'}
         />
+        <ExportTrack/>
         <Header>Liste des positions VMS affichées</Header>
         <PositionsTable openBox />
       </TrackRequestBody>
@@ -118,10 +129,10 @@ const TrackRequestButton = styled(MapComponentStyle)`
   border-radius: 1px;
   cursor: pointer;
   height: 30px;
-  margin-right: 505px;
-  opacity: 1;
+  margin-right: ${props => props.sidebarIsOpen ? 505 : -45}px;
+  right: ${props => props.isRightMenuOpen && props.sidebarIsOpen ? 55 : 10}px;
+  opacity: ${props => props.sidebarIsOpen ? 1 : 0};
   position: absolute;
-  right: ${p => p.isRightMenuOpen ? 55 : 10}px;
   top: 153px;
   transition: all 0.5s, right 0.3s;
   width: 30px;
@@ -129,17 +140,15 @@ const TrackRequestButton = styled(MapComponentStyle)`
 `
 
 const TrackRequestBody = styled(MapComponentStyle)`
-  animation: ${p => p.isRightMenuOpen && p.isTrackRequestOpen
-    ? 'vessel-box-opening-with-right-menu-hover'
-    : 'vessel-box-closing-with-right-menu-hover'} 0.3s ease forwards;
   background: ${COLORS.background};
   border-radius: 2px;
   color: ${COLORS.slateGray};
   font-size: 13px;
+  text-align: left;
   margin-right: ${p => p.isTrackRequestOpen ? '540px' : '217px'};
   opacity: ${p => p.isTrackRequestOpen ? '1' : '0'};
   position: absolute;
-  right: 10px;
+  right: ${props => props.isRightMenuOpen && props.sidebarIsOpen ? 55 : 10}px;
   top: 118px;
   transition: all 0.3s;
   visibility: ${p => p.isTrackRequestOpen ? 'visible' : 'hidden'};
