+++ frontend/src/features/vessel_sidebar/actions/track_request/exportTrackOptions.js
@@ -1,0 +1,62 @@
+import { ExportToCsv } from 'export-to-csv'
+
+const optionsCSV = {
+  fieldSeparator: ',',
+  quoteStrings: '"',
+  decimalSeparator: '.',
+  showLabels: true,
+  showTitle: false,
+  useTextFile: false,
+  useBom: true,
+  useKeysAsHeaders: true
+}
+
+export const csvExporter = new ExportToCsv(optionsCSV)
+
+// These properties are ordered for the CSV column order
+export const csvOrder = {
+  vesselName: {
+    code: 'vesselName',
+    name: 'Nom'
+  },
+  externalReferenceNumber: {
+    code: 'externalReferenceNumber',
+    name: 'Marq. Ext.'
+  },
+  ircs: {
+    code: 'ircs',
+    name: 'C/S'
+  },
+  mmsi: {
+    code: 'mmsi',
+    name: 'MMSI'
+  },
+  internalReferenceNumber: {
+    code: 'internalReferenceNumber',
+    name: 'CFR'
+  },
+  flagState: {
+    code: 'flagState',
+    name: 'Pavillon'
+  },
+  dateTime: {
+    code: 'dateTime',
+    name: 'GDH (UTC)'
+  },
+  latitude: {
+    code: 'latitude',
+    name: 'Latitude'
+  },
+  longitude: {
+    code: 'longitude',
+    name: 'Longitude'
+  },
+  course: {
+    code: 'course',
+    name: 'Cap'
+  },
+  speed: {
+    code: 'speed',
+    name: 'Vitesse'
+  }
+}
